ChatBot/requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langchain-core
faiss-cpu  
langsmith
python-dotenv
beautifulsoup4
unstructured
